+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## v0.17.2 - 2026-07-21
+
+- Added tracking events to `DatasourceConfigWizard` component
+- Export provisioning scripts from dsconfig/schema
+
 ## v0.17.1 - 2026-07-16
 
 - Added new `DatasourceConfigWizard` component ( schema driven config page )